Fixed all comments

+++ app/main.py
@@ -14,22 +14,19 @@ def move_file(command: str) -> None:
         directory, file_name = os.path.split(new_file_path)
     except TypeError:
         raise TypeError(f"{new_file_path} is not string")
+    if file_name == old_file_path:
+        raise ValueError("The source and destination represent "
+                         "the same file.")
     if len(directory) == 0:
-        try:
-            shutil.copy(old_file_path, file_name)
-            os.remove(old_file_path)
-        except shutil.SameFileError:
-            raise shutil.SameFileError("The source and destination represent "
-                                       "the same file.")
+        with open(old_file_path, "r") as old_file:
+            with open(new_file_path, "w") as new_file:
+                new_file.write(old_file.read())
+        os.remove(old_file_path)
     else:
         new_file_path = os.path.join(directory, file_name)
         new_directory = os.path.dirname(new_file_path)
         if not os.path.exists(new_directory):
-            try:
-                os.makedirs(new_directory)
-            except FileExistsError:
-                raise FileExistsError(f"The file cannot be created because "
-                                      f"it already exists: {new_directory}")
+            os.makedirs(new_directory, exist_ok=True)
         try:
             shutil.copy(old_file_path, new_file_path)
             os.remove(old_file_path)
